feat(mes): check for mTLS and connected zoneEgress + warning
