+++ src/app/services/locales/en-us/index.yaml
@@ -46,4 +46,9 @@ services:
   mesh-external-service:
     href:
       docs: '{KUMA_DOCS_URL}/networking/meshexternalservice/?{KUMA_UTM_QUERY_PARAMS}'
+    notifications:
+      mtls-warning: !!text/markdown |
+        mTLS is not enabled on this mesh. <a href="{KUMA_DOCS_URL}/policies/mutual-tls/?{KUMA_UTM_QUERY_PARAMS}" target="_blank">Enable mTLS for MeshExternalService to work</a>
+      no-zone-egress: !!text/markdown |
+        There is no connected ZoneEgress in this mesh. <a href="{KUMA_DOCS_URL}/production/cp-deployment/zoneegress?{KUMA_UTM_QUERY_PARAMS}" target="_blank">Add a ZoneEgress for MeshExternalService to work</a>
